+++ src/Http/Controllers/RelationModelFieldController.php
@@ -4,11 +4,27 @@ declare(strict_types=1);
 
 namespace MoonShine\Http\Controllers;
 
+use App\Models\PolyComment;
+use Illuminate\Contracts\Auth\Guard;
+use Illuminate\Contracts\Auth\StatefulGuard;
 use Illuminate\Contracts\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\MorphOneOrMany;
+use MoonShine\ActionButtons\ActionButton;
+use MoonShine\Collections\MoonShineRenderElements;
+use MoonShine\Components\FormBuilder;
 use MoonShine\Components\TableBuilder;
 use MoonShine\Contracts\Fields\Relationships\HasAsyncSearch;
+use MoonShine\Enums\JsEvent;
+use MoonShine\Fields\Field;
+use MoonShine\Fields\Fields;
+use MoonShine\Fields\Hidden;
+use MoonShine\Fields\Relationships\HasMany;
+use MoonShine\Fields\Relationships\ModelRelationField;
 use MoonShine\Fields\Relationships\MorphTo;
 use MoonShine\Http\Requests\Relations\RelationModelFieldRequest;
+use MoonShine\Resources\ModelResource;
+use MoonShine\Support\AlpineJs;
 use MoonShine\Support\DBOperators;
 use MoonShine\Support\TableRowRenderer;
 use Symfony\Component\HttpFoundation\Response;
@@ -112,5 +128,107 @@ class RelationModelFieldController extends MoonShineController
         }
 
         return $value;
+    }
+
+    /**
+     * @throws Throwable
+     */
+    public function hasManyForm(RelationModelFieldRequest $request): string
+    {
+        $parent = $request->getResource()?->getItemOrInstance();
+
+        /** @var HasMany $field */
+        $field = $request->getField();
+
+        /** @var ModelResource $resource */
+        $resource = $field->getResource();
+        $item = $field->getResource()
+            ->setItemID($request->get('_key'))
+            ->getItemOrInstance();
+        $update = $item->exists;
+        $relation = $parent?->{$field->getRelationName()}();
+
+        $field->resolveFill($parent->toArray(), $parent);
+
+        $action = $update
+            ? static fn (Model $data) => $resource->route('crud.update', $data->getKey())
+            : static fn (?Model $data) => $resource->route('crud.store');
+
+        $isAsync = $resource->isAsync() || $field->isAsync();
+
+        $getFields = function () use ($resource, $field, $isAsync, $parent, $update) {
+            $fields = $resource->getFormFields();
+
+            $fields->onlyFields()
+                ->each(fn (Field $nestedFields): Field => $nestedFields->setParent($field))
+                // Uncomment if you need a parent resource
+                //->onlyRelationFields()
+                //->each(fn (ModelRelationField $nestedFields): Field => $nestedFields->setParentResource($resource))
+            ;
+
+            return $fields->when(
+                $field->getRelation() instanceof MorphOneOrMany,
+                fn (Fields $f) => $f->push(
+                    Hidden::make($field->getRelation()?->getMorphType())
+                        ->setValue($parent::class)
+                )
+            )->when(
+                $update,
+                fn (Fields $f) => $f->push(
+                    Hidden::make('_method')->setValue('PUT'),
+                )
+            )
+                ->push(
+                    Hidden::make($field->getRelation()?->getForeignKeyName())
+                        ->setValue($parent->getKey())
+                )
+                ->push(Hidden::make('_async_field')->setValue($isAsync))
+                ->toArray();
+        };
+
+        return (string) FormBuilder::make($action($item))
+            ->fields($getFields)
+            ->reactiveUrl(
+                fn (): string => moonshineRouter()
+                    ->reactive(key: $item?->getKey(), page: $resource->formPage(), resource: $resource)
+            )
+            ->name($resource->uriKey())
+            ->switchFormMode(
+                $isAsync,
+                [
+                    $resource->listEventName($field->getRelationName()),
+                    AlpineJs::event(JsEvent::FORM_RESET, $resource->uriKey()),
+                ]
+            )
+            ->when(
+                $update,
+                fn (FormBuilder $form): FormBuilder => $form->fillCast(
+                    $item,
+                    $resource->getModelCast()
+                ),
+                fn (FormBuilder $form): FormBuilder => $form->fillCast(
+                    array_filter([
+                        $field->getRelation()?->getForeignKeyName() => $parent?->getKey(),
+                        ...$field->getRelation() instanceof MorphOneOrMany
+                            ? [$field->getRelation()?->getMorphType() => $parent?->getMorphClass()]
+                            : [],
+                    ], static fn ($value) => filled($value)),
+                    $resource->getModelCast()
+                )
+            )
+            ->submit(__('moonshine::ui.save'), ['class' => 'btn-primary btn-lg'])
+            ->onBeforeFieldsRender(fn (Fields $fields): MoonShineRenderElements => $fields->exceptElements(
+                fn (mixed $field): bool => $field instanceof ModelRelationField
+                    && $field->toOne()
+                    && $field->column() === $relation->getForeignKeyName()
+            ))
+            ->buttons($resource->getFormButtons())
+            ->redirect(
+                $isAsync ?
+                    null
+                    : moonshineRequest()
+                    ->getResource()
+                    ?->formPageUrl($parent)
+            );
     }
 }
